Fix data import
