+++ packages/beacon-node/test/utils/network.ts
@@ -1,7 +1,5 @@
 import {PeerId} from "@libp2p/interface";
 import {createSecp256k1PeerId} from "@libp2p/peer-id-factory";
-import {BitArray} from "@chainsafe/ssz";
-import {ATTESTATION_SUBNET_COUNT, SYNC_COMMITTEE_SUBNET_COUNT} from "@lodestar/params";
 import {ssz} from "@lodestar/types";
 import {ChainForkConfig, createBeaconConfig} from "@lodestar/config";
 import {
@@ -163,26 +161,4 @@ export function onPeerConnect(network: Network): Promise<void> {
 
 export function onPeerDisconnect(network: Network): Promise<void> {
   return new Promise<void>((resolve) => network.events.on(NetworkEvent.peerDisconnected, () => resolve()));
-}
-
-/**
- * Generate valid filled attnets BitVector
- */
-export function getAttnets(subnetIds: number[] = []): BitArray {
-  const attnets = BitArray.fromBitLen(ATTESTATION_SUBNET_COUNT);
-  for (const subnetId of subnetIds) {
-    attnets.set(subnetId, true);
-  }
-  return attnets;
-}
-
-/**
- * Generate valid filled syncnets BitVector
- */
-export function getSyncnets(subnetIds: number[] = []): BitArray {
-  const syncnets = BitArray.fromBitLen(SYNC_COMMITTEE_SUBNET_COUNT);
-  for (const subnetId of subnetIds) {
-    syncnets.set(subnetId, true);
-  }
-  return syncnets;
 }
