+++ packages/beacon-node/test/utils/networkData.ts
@@ -1,0 +1,24 @@
+import {BitArray} from "@chainsafe/ssz";
+import {ATTESTATION_SUBNET_COUNT, SYNC_COMMITTEE_SUBNET_COUNT} from "@lodestar/params";
+
+/**
+ * Generate valid filled attnets BitVector
+ */
+export function getAttnets(subnetIds: number[] = []): BitArray {
+  const attnets = BitArray.fromBitLen(ATTESTATION_SUBNET_COUNT);
+  for (const subnetId of subnetIds) {
+    attnets.set(subnetId, true);
+  }
+  return attnets;
+}
+
+/**
+ * Generate valid filled syncnets BitVector
+ */
+export function getSyncnets(subnetIds: number[] = []): BitArray {
+  const syncnets = BitArray.fromBitLen(SYNC_COMMITTEE_SUBNET_COUNT);
+  for (const subnetId of subnetIds) {
+    syncnets.set(subnetId, true);
+  }
+  return syncnets;
+}
